requirements.txt
aiosmtpd==1.4.2
asyncio==3.4.3
autopep8==2.0.2
awscrt==0.16.9
babel==2.12.1
black==23.3.0
boto==2.49.0
boto3==1.26.123
brotli==1.0.9
brotlipy==0.7.0
coverage==7.2.1
cryptography==38.0.4
debugpy==1.6.6
dj-database-url==2.0.0
django-allauth==0.52.0
django-compressor==4.3.1
django-crispy-forms==1.14.0
django-htmx==1.14.0
django-nose==1.4.7
django-storages==1.13.2
django-widget-tweaks==1.4.12
djlint==1.25.0
factory-boy==3.2.1
flake8==6.0.0
gunicorn==20.1.0
ipython==8.13.1
markdown==3.4.3
pillow==9.4.0
pip-chill==1.0.3
pytest-django==4.5.2
python-dotenv==1.0.0
python-memcached==1.59
pytz==2023.3
toml==0.10.2
types-cryptography==3.3.23
whitenoise==6.4.0
psycopg